+++ CHANGELOG.md
@@ -3,7 +3,9 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ## Changed
+- minor fixes in documentation typos and argument names.
 - `Domain` is moved from `hypertunity.optimisation` to the `hypertunity` package.
+- rename `TableReporter` to `Table` and `TensorboardReporter` to `Tensorboard`.
 
 ## [0.4.0] - 2019-09-15
 ## Added